+++ include/rocksdb/compaction_filter.h
@@ -39,6 +39,8 @@ class CompactionFilter {
     // Is this compaction requested by the client (true),
     // or is it occurring as an automatic compaction process
     bool is_manual_compaction;
+    // Which column family this compaction is for.
+    uint32_t column_family_id;
   };
 
   virtual ~CompactionFilter() {}